requirements.txt
flask>=2.0.1
flask-socketio>=5.1.1
pyserial>=3.5
mysql-connector-python>=8.0.26
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.0.0
joblib>=1.0.1
python-engineio>=4.2.1
python-socketio>=5.4.0
matplotlib
seaborn 